// services/sdk/node/src/generated/openapi.ts
export interface components {
    schemas: {
// ...
        Problem: {
            type: string;
            title: string;
            status: number;
            code: string;
            detail?: string;
            request_id?: string;
            retry_after?: number;
            errors?: {
                [key: string]: string[];
            };
        };
// ...
    };
// ...
}